+++ commons/src/lib.rs
@@ -4,4 +4,3 @@ pub mod gem_version;
 pub mod gemfile_lock;
 pub mod layer;
 pub mod metadata_digest;
-pub mod output;